requirements.txt
Django==3.2.15
#djangorestframework==3.13.1
djangorestframework==3.12.4
djaodjin-deployutils==0.10.3
djaodjin-extended-templates==0.3.4
djaodjin-pages==0.6.7
djaodjin-survey==0.8.0
gunicorn==19.9.0
jinja2==3.1.2
html5lib==1.1
monotonic==1.6
openpyxl==3.0.9
python-pptx==0.6.21
pytz==2022.1

# These have C/C++ components, they can be installed through the recent wheel
# pip feature or through the native package manager
cairocffi==1.3.0
coverage==6.3.2
cryptography==38.0.3
psycopg2-binary==2.9.3
setproctitle==1.2.3

# We cannot build the assets (i.e. webpack-conf-paths.json) unless DEBUG=1
# and django-extensions is installed.
django-extensions==3.2.0          # 3.2.0 required by Django==4.0

# To generate API docs
coreapi==2.3.3
drf-yasg==1.8.0
Pygments==2.12.0
uritemplate==4.1.1
